Bug fixed - Random collors/Objects level
The color ends up being defined by the ArtType present in the LevelPieceBasedSetup, and in this case the choice of the setup was being made via the _index variable that started at 1:

So when the instantiation actually started the index changed to 2 and that index is used.

In this case, when there is a restart, the entire scene is reloaded, which causes this index to be reset to 1 and the value 2 to be used again, always keeping the same color.

So what was done is the randomization of the index in awake so that the initial color is always randomized:
  private void Awake()
     {
         _index=Random.Range(0, levelPieceBasedSetups.Count);
     }

That way the level initialized at restart is always a different one.