Bug fixed after hide player
As the PlayerController class is a child of the Singleton class, methods it declares with the same name as the parent class's methods may end up overriding the parent class's methods. So what is happening is exactly this:
As the Awake method of the Singleton class is private, the compiler doesn't even report anything, as there is no compilation error, but what ends up happening is that when Unity calls the Awake method of the PlayerController class, it only calls its Awake and not Singleton's, which causes the Instance variable to become null.
What we have to do then is first pass the Awake method of the Singleton class to protected, so that it can be accessed by the child classes.
This helps us because now the IDEs tend to show with a green underline that there is some detail there, not necessarily an error, but something that is good to pay attention to.
In this case, it already gives us the message that this Awake method of the child class hides the Awake method of the parent class, that is, whoever searches for the Awake method of our PlayerController class will only find that of the child class:
Being protected we can now call the Awake method of the parent class as well:
And since we have already “solved” the problem of the fact that the awake method of the parent class is being called, we can put the new keyword before the method declaration:
This only serves to remove the green underline, as it makes it clear to the compiler that we are actually hiding the method of the parent class, but that is the point.
As a result, all classes that take the PlayerController reference should work correctly: